Routes: Rename Mentorship Fa'24 Routes
TEST=manual
