+++ js/src/patina/pages/mentor-fa24/Mentor.page.tsx
@@ -1,11 +1,11 @@
 import { Button, Center, Space, Text, useMantineTheme } from '@mantine/core'
 import { Hero } from '@/patina/components/Hero'
 // import { Impact } from '@/patina/pages/mentor/impact/Impact'
-import { About } from '@/patina/pages/mentor/about/About'
+import { About } from '@/patina/pages/mentor-fa24/about/About'
 import { ContentPage } from '@/patina/components/ContentPage.tsx'
 import { imageUrls } from '@/patina/assets/images'
 import styles from './Mentor.module.css'
-import { ProgramDetails } from '@/patina/pages/mentor/program/ProgramDetails.tsx'
+import { ProgramDetails } from '@/patina/pages/mentor-fa24/program/ProgramDetails.tsx'
 
 const mentorGoogleFormUrl =
   'https://docs.google.com/forms/d/1yWPT7gYtpiKbBO6IKgcTygTzuHdAZaSomgu0BnDW23s/viewform'
